in structure HeadPath int swaped for short - saving memory possible due to vector<HeadPath> is reaching big size during play, Empty vector mechanism should be implemented because we do not need to store whole head path tracking - we just need last places equals to how many tails we have got

+++ Unit1.cpp
@@ -14,7 +14,7 @@ TForm1 *Form1;
 //////////////////////////////////////////////////////////////////////////////
 struct TheadPath
 {
- int x,y;
+ short int x,y;
 
  TheadPath(int a, int b): x(a), y(b) {}
 };
